Less strict parsing of patchDates mapping config (#2271)
The mapping required individual fields to be separated by not just by a comma, but by comma with whitespace on both sides. As a result, even the example given in the [documentation](https://retorque.re/zotero-better-bibtex/installation/preferences/hidden-preferences/#patchdates)

```
date-added = dateAdded, timestamp=dateModified
```

did not work. This fix changes the regular expression so that whitespace before and after the comma is optional.