Fix syntax for links to citing page.
Links to citing not generating correctly. I'm not very familiar with whatever SSG you're using (looks like Hugo?), but I looked another page with working links and copied the code from there. If it's not the right code in the PR, at least it will show you where the changes are needed. Thanks for a great plugin!